[cord-12]bootstrap

+++ cord19-explorer/src/main/java/edu/asu/diging/cord19/explorer/core/mongo/PublicationSearchProvider.java
@@ -8,6 +8,6 @@ public interface PublicationSearchProvider {
 
     long searchResultSize(String title);
 
-    List<PublicationImpl> getRequestedPage(String title, Integer currentPage, Integer size);
+    List<PublicationImpl> getRequestedPage(String title, Long currentPage, Integer size);
 
 }
